Version1.8 debug刪除留言、使用者介面留言順序

+++ proTest/WebApplication2/Models/DBmanager.cs
@@ -37,7 +37,7 @@ namespace WebApplication2.Models
                         userName = reader.GetString(reader.GetOrdinal("userName")),
                         main = reader.GetString(reader.GetOrdinal("main")),
                         date = reader.GetDateTime(reader.GetOrdinal("date")),
-                        score = reader.GetDouble(reader.GetOrdinal("score"))
+                        score = reader.GetInt32(reader.GetOrdinal("score"))
                     };
                     m.Add(account);
                 }
